MP-168 :wrench: Chore: 잡다한 과업 수행
* UlidIdGenerator의 상위 객체를 BeforeExecutionGenerator에서 Assigned로 변경
* convertFileSrcToBinaryData의 반환 타입을 JsonNode에서 ArrayNode로 변경
* 기타 불필요한 로직 제거 등 간소화

+++ src/main/java/kr/modusplant/domains/common/app/service/MultipartDataProcessor.java
@@ -18,8 +18,7 @@ import java.nio.charset.StandardCharsets;
 import java.util.Base64;
 import java.util.List;
 
-import static kr.modusplant.global.vo.CamelCaseWord.DATA;
-import static kr.modusplant.global.vo.CamelCaseWord.ORDER;
+import static kr.modusplant.global.vo.CamelCaseWord.*;
 import static kr.modusplant.global.vo.FileSystem.FILENAME;
 import static kr.modusplant.global.vo.FileSystem.SRC;
 
@@ -32,11 +31,11 @@ public class MultipartDataProcessor {
     private static final UlidIdGenerator generator = new UlidIdGenerator();
 
     public JsonNode saveFilesAndGenerateContentJson(PostType postType, List<MultipartFile> parts) throws IOException {
-        String fileUlid = generator.generate(null,null,null, EventType.INSERT);
+        String fileUlid = generator.generate(null, null, null, EventType.INSERT);
         ArrayNode contentArray = objectMapper.createArrayNode();
-        int order=1;
+        int order = 1;
         for (MultipartFile part:parts) {
-            contentArray.add(convertSinglePartToJson(postType,fileUlid, part,order++));
+            contentArray.add(convertSinglePartToJson(postType, fileUlid, part, order++));
         }
         return contentArray;
     }
@@ -46,21 +45,21 @@ public class MultipartDataProcessor {
         String filename = part.getOriginalFilename();
 
         ObjectNode node = objectMapper.createObjectNode();
-        node.put(FILENAME,filename);
-        node.put(ORDER,order);
+        node.put(FILENAME, filename);
+        node.put(ORDER, order);
 
         FileType fileType = FileType.from(contentType);
         if (fileType == FileType.TEXT) {
             String text = new String(part.getBytes(), StandardCharsets.UTF_8);
-            node.put("type",fileType.getValue());
-            node.put(DATA,text);
+            node.put(TYPE, fileType.getValue());
+            node.put(DATA, text);
         } else if (fileType.getUploadable()) {
-            String fileKey = generateFileKey(postType,fileUlid,fileType,filename,order);
-            s3FileService.uploadFile(part,fileKey);
-            node.put("type",fileType.getValue());
-            node.put(SRC,fileKey);
+            String fileKey = generateFileKey(postType, fileUlid, fileType, filename, order);
+            s3FileService.uploadFile(part, fileKey);
+            node.put(TYPE, fileType.getValue());
+            node.put(SRC, fileKey);
         } else {
-            throw new IllegalArgumentException("Unsupported file type: "+contentType);
+            throw new IllegalArgumentException("지원되지 않는 파일 타입입니다.");
         }
         return node;
     }
@@ -79,15 +78,15 @@ public class MultipartDataProcessor {
         return directory + filename;
     }
 
-    public JsonNode convertFileSrcToBinaryData(JsonNode content) throws IOException {
+    public ArrayNode convertFileSrcToBinaryData(JsonNode content) throws IOException {
         ArrayNode newArray = objectMapper.createArrayNode();
-        for(JsonNode node:content) {
+        for (JsonNode node : content) {
             ObjectNode objectNode = node.deepCopy();
-            if(node.isObject() && node.has(SRC)) {
+            if (node.isObject() && node.has(SRC)) {
                 String src = objectNode.get(SRC).asText();
                 byte[] fileBytes = s3FileService.downloadFile(src);
                 String base64Encoded = Base64.getEncoder().encodeToString(fileBytes);
-                objectNode.put(DATA,base64Encoded);
+                objectNode.put(DATA, base64Encoded);
                 objectNode.remove(SRC);
             }
             newArray.add(objectNode);
@@ -95,13 +94,11 @@ public class MultipartDataProcessor {
         return newArray;
     }
 
-    public void deleteFiles(JsonNode content) throws IOException {
+    public void deleteFiles(JsonNode content) {
         for (JsonNode node : content) {
-            if (node.isObject()) {
-                if (node.has(SRC)) {
-                    String src = node.get(SRC).asText();
-                    s3FileService.deleteFiles(src);
-                }
+            if (node.isObject() && node.has(SRC)) {
+                String src = node.get(SRC).asText();
+                s3FileService.deleteFiles(src);
             }
         }
     }
